Move combine helpers to a new file
